Lib: Move addQueryArgs to route module
