+++ client/lib/route/test/index.js
@@ -17,6 +17,57 @@ describe( 'route', function() {
 		route = require( '../' );
 	} );
 
+	describe( '#addQueryArgs()', () => {
+		it( 'should error when args is not an object', () => {
+			var types = [
+				undefined,
+				1,
+				true,
+				[],
+				'test',
+				function() {}
+			];
+
+			types.forEach( type => {
+				expect( () => {
+					route.addQueryArgs( type );
+				} ).to.throw( Error );
+			} );
+		} );
+
+		it( 'should error when url is not a string', () => {
+			var types = [
+				{},
+				undefined,
+				1,
+				true,
+				[],
+				function() {}
+			];
+
+			types.forEach( type => {
+				expect( () => {
+					route.addQueryArgs( {}, type );
+				} ).to.throw( Error );
+			} );
+		} );
+
+		it( 'should return same URL with ending slash if passed empty object for args', () => {
+			const url = route.addQueryArgs( {}, 'https://wordpress.com' );
+			expect( url ).to.eql( 'https://wordpress.com/' );
+		} );
+
+		it( 'should add query args when URL has no args', () => {
+			const url = route.addQueryArgs( { foo: 'bar' }, 'https://wordpress.com' );
+			expect( url ).to.eql( 'https://wordpress.com/?foo=bar' );
+		} );
+
+		it( 'should persist existing query args and add new args', () => {
+			const url = route.addQueryArgs( { foo: 'bar' }, 'https://wordpress.com?search=test' );
+			expect( url ).to.eql( 'https://wordpress.com/?search=test&foo=bar' );
+		} );
+	} );
+
 	describe( 'getSiteFragment', function() {
 		describe( 'for the root path', function() {
 			it( 'should return false', function() {
